Refactor dataset and loss head modules
