+++ sesemi/datasets/__init__.py
@@ -1,0 +1,12 @@
+from . import (
+    base,
+    cifar,
+    class_weighted,
+    concat,
+    image_file,
+    image_folder,
+    mnist,
+    pseudo,
+    stl10,
+)
+from .base import register_dataset, dataset
