+++ sesemi/models/backbones/__init__.py
@@ -3,3 +3,7 @@
 # =============================================#
 """Backbone models."""
 from . import base, resnet, timm, torchvision
+from .base import Backbone
+from .resnet import CIFARResNet
+from .timm import PyTorchImageModels
+from .torchvision import TorchVisionBackbone
