+++ sesemi/models/heads/__init__.py
@@ -2,3 +2,5 @@
 # Copyright 2021, Flyreel. All Rights Reserved.
 # =============================================#
 """Head models."""
+from .base import Head
+from .linear import LinearHead
