+++ sesemi/models/loss_heads/__init__.py
@@ -1,0 +1,6 @@
+from .base import LossHead
+from .consistency import ConsistencyLossHead, EMAConsistencyLossHead
+from .entropy_minimization import EntropyMinimizationLossHead
+from .fixmatch import FixMatchLossHead
+from .jigsaw_prediction import JigsawPredictionLossHead
+from .rotation_prediction import RotationPredictionLossHead
