+++ sesemi/trainer/conf/base/jigsaw_prediction/model.yaml
@@ -5,6 +5,6 @@ learner:
       regularization_loss_heads:
         jigsaw_prediction:
           head:
-            _target_: sesemi.models.heads.loss.JigsawPredictionLossHead
+            _target_: sesemi.loss_heads.loss.JigsawPredictionLossHead
             data: unsupervised
             backbone: supervised_backbone
