+++ sesemi/trainer/conf/base/rotation_entropy_minimization/model.yaml
@@ -5,7 +5,7 @@ learner:
       regularization_loss_heads:
         rotation_prediction:
           head:
-            _target_: sesemi.models.heads.loss.RotationPredictionLossHead
+            _target_: sesemi.models.loss_heads.RotationPredictionLossHead
             data: unsupervised
             backbone: supervised_backbone
           scheduler:
@@ -15,7 +15,7 @@ learner:
           scale_factor: 1.0
         entropy_minimization:
           head:
-            _target_: sesemi.models.heads.loss.EntropyMinimizationLossHead
+            _target_: sesemi.models.loss_heads.EntropyMinimizationLossHead
             data: unsupervised
             backbone: supervised_backbone
             head: supervised_head
