+++ sesemi/trainer/conf/base/rotation_prediction/model.yaml
@@ -5,6 +5,6 @@ learner:
       regularization_loss_heads:
         rotation_prediction:
           head:
-            _target_: sesemi.models.heads.loss.RotationPredictionLossHead
+            _target_: sesemi.models.loss_heads.RotationPredictionLossHead
             data: rotation_prediction
             backbone: supervised_backbone
